Wrong buffer size in DefaultDisplaySpec
We are using sizeof(pointer) as size of the buffer. This is wrong as size of pointer
is not equal to size of array it points to. Size of pointer is usually
4 bytes on 32 bit system or size of the address register on the particular
machine controls the size of the pointer. In this case we already know the
size as we are allocating the char array on heap. This patch makes necessary
changes in this direction.