Refactor InputPointer.
Currently, WaylandDisplay initializes cursors and inputdevice
updates it as needed. There are certain dependencies on inputdevice
for the cursor to create its surface buffer. This has led to un-necessary
apis in inputdevice. The patch changes so that all the functionality
of Cursor is encapsulated in WaylandCursor class and it is owned by
input device. After this change WaylandDisplay knows nothing about
cursor and how it's updated. With this change we also move all
global declarations to wayland_global