Simplify retrieveing primary screen attributes.
In DefaultDisplaySpec we retrieve the first screen in screen list, check
all available modes. The attibutes of first mode with WL_OUTPUT_MODE_CURRENT
enabled are retrieved as geometry of screen. The patch introduces
the concept of primary screen (output). The api is part of WaylandDisplay
and logic to decide which is primary screen is internal to WaylandDisplay.
Currently, first screen in screen_list is always primary screen.
We dont have to store all modes of screen but only last mode with
WL_OUTPUT_MODE_CURRENT flag enabled.