Refactor window and move event flushing to separate thread.
Currently, window is responsible for creating shell surface,
egl window and also needs to manage it's different states.
Also, with multi process model we would need it to be
responsible for part of those duties depending on current process.

The patch does the follwing:
1)Moves Shell surface related things to shell_surface class.
This should be responsible to update the shell surface state
and managing its size.
2)Move EGL window related things to egl_window class.
3)Re-work the internals of window class and introduce necessary
apis for RootWindowHost to use.

After these changes, EGLWindow is not created till RealizeAcceleratedWidget
is not called.
Now, reading and flushing display data is done on a separate thread.
Ideally we should be able to move the polling also to this thread.