EventFactoryWayland should observe message loop destruction.
EventFactoryWayland registers itself for observing
any processed tasks but is never removed from the list.
Expectation here is that it would be cleaned up during
exit phase. Correct solution would be to make
EventFactoryWayland register as destruction observer
of the message loop.

Now, EventFactoryWayland adds itself to destruction
observer list and does proper de-registration when
the message loop is about to be destroyed.