Introduce WaylandShellSurface class.
The patch introduces WaylandShellSurface class.
It's a simple class which would create a shell surface and
resposible for all it's activity. Currently, only support
for TOPLEVEL is added. Support for other functionality
i.e Min, Maximize, Popup, Transiant etc should be added
as we support other use cases. The main consumer of this
class is Window.