Add WaylandDispatcher class.
We currently handle all wayland
flushing in ui thread. It would be nice to
offload it to a worker thread. WaylandDispatcher
class is responsible for reading any pending events
and flush wayland display in a worker thread.
It has apis which could be used to post task to
be handled by worker thread or main thread.