impl: Fix the spec_ and its size allocation
"2 * sizeof bounds.width()" for the size of spec_ is very wrong and, besides,
the format also needed to be changed for the correct one (something like
"%dx%d*2").

This changes here explicitly exposes the problem of OutputHandleMode not coming
in time during Chromium's initialization. But I'll be fixing it completely in
the upcoming commit.