impl: OzoneDisplay::OnChannelEstablished shouldn't send old data.
Once the channel is established, the browser process already has been
running for awhile and the output geometries in question were already
accessed by the UI. So it's just useless to dispatch a late
OutputSizeChanged.

Signed-off-by: Tiago Vignatti <tiago.vignatti@intel.com>

The patch was imported from https://github.com/tiagovignatti/ozone-wayland/
as of commit id ffd065d.

Rebased.

Signed-off-by: Kalyan Kondapally <kalyan.kondapally@intel.com>