patches: Add patch for supporting Chrome browser
We're already trying to upstream this patch:

https://codereview.chromium.org/36953002/

To get these changes upstreamed, there are a couple of changes that needs to be
tackled before:
- add class DesktopRootWindowHostOzone
- move desktop support into ui/base so we don't reference views from ozone
  platform code.
- change the namespace ozonewayland to views in most of the impl/desktop_* so
  we can explicit declare dependency on views for the classes on those files

This is a small patch and, although the solution is a hack (per comments in the
CL), the maintenance of it in here doesn't bring a big burden to us. So I think
it's worth to keep it here and have the chrome browser target successfully
compiling and working.