Avoid duplicate route id in Gpu Process.
GpuChannelManager generates unique routeid for every new
imagetransportsurface. On our side, we register a routeid
for our display channel. Depending on our routeid, we will
conflict with the one generated by GpuChannelManager. To avoid
this, now we hardcore our routeid with a negative sign to be
unique from the ones generated by GpuChannelManager. This is
one of the main reason for random crashes. This also fixes
the crash in multi window case.