Delete associated window when DRWH is destroyed.
This is in preparation for supporting multi window mode.
Currently, we always assume single root window and it is destroyed
when browser is closed. We would eventually want to support multi
window mode in the form of popups or additional top windows. This
patch adds support to delete the window once DRWH is closed.