Fix Seg fault with Content shell.
With recent changes, window doesn't set the shelltype but
expects the caller to do it. In DRWH, we don't yet handle
the case when Root Window is of type Popup or Tooltip. This
patch makes changes to treat them similar to Menu for now.
Also, the patch removes assumption of popup having a parent
as it doesn't seem the case with content shell.Instead we use
g_current_capture to track if menu or popup is open. This
also removes another un-necessary book kepping.