Keyboard: Don't create a new key event in DRWH.
DRWH is responsible for dispatching key events to delegate.
Instead of using the current event, we construct a new key event.
This seems unnecessary and also any character information is not
copied in this case(causing unnecessary re-calculations). This
CL changes so that, we re-use existing event instead of creating
new one.