DRWH: Previous bounds is not always correct.
We save previous bounds during the construction phase and it's not
updated as and when bounds is really changed. (i.e. SetSize or
SetBounds call). Right way to track the restore/previous bounds is
to store the value before window state is changed to FullScreen,
Maximized etc. This CL does changes in this direction