Wayland: Pass wl_display instead of display fd to Dispatcher.
Dispatcher is responsible for polling on the display fd and
flushing wayland events. Dispatcher takes fd and queries wl_display
from WaylandDisplay every time. It's bit confusing from the api's
as to how Dispatcher maps fd with wl_display (to flush etc). It would
make it clean, if wl_display is passed during construction phase
instead of fd. This CL makes necessary changes in this direction.