OzoneDisplayChannelHost: Add support for observing ChildProcessHost.
This CL adds necessary support in OzoneDisplayChannelHost to observe for new
child process host connections. We rely on this to establish a Channel Connection.
We already have similar support in OzoneChildProcessObserver class, but it seems
un-necessary as this information is very much internal to OzoneDisplayChannelHost.
OzoneChildProcessObserver will be removed in a follow up patch.