WaylandWindow: Replace SetBounds API with Resize.
We never set position of WaylandWindow but only the size attributes
(i.e. Width and Height). However, SetBounds accepts Rect as the function
parameter and we end up passing redundant information (i.e. Position) which
is always ignored. This CL replaces SetBounds API with an Resize API which
accepts only Width and Height.