DRWH: Explicitly restore Window state when exiting from FullScreen mode.
Currently, if shell (in WaylandWindow) is of type FullScreen, we set it to
TopLevel and vice versa. However, we also need to deal with various
combinations like Maximize->FullScreen->Maximize->Normal, a frameless window
in future etc. This CL changes it so that WaylandWindow doesn't make any
assumptions and all state changes need to be explicitly set.