ChannelHost: Set ChannelHost as handler for control messages with BGCHFactor
OzoneDisplayChannelHost, sets itself as filter with GpuProcessHost and
listens to any incoming messages from GPU process. This approach works fine
but we seem to be using the wrong class to do this.

BrowserGPUChannelHostFactory provides the needed API interface. This CL
makes changes to replace GPUProcessHost and IPC::ChannelProxy::MessageFilter
usage with the API provided by BGCHFactory. Most of the changes are based
on similar functionality provided by BrowserCompositorOutputSurfaceProxy.

Fixes #165 Fixes #24