WindowStateHandler: Introduce RemoteStateChangeHandler.
OzoneDisplayChannelHost handles both sending and receiving of IPC messages
from GPU Process. This CL introduces RemoteStateChangeHandler which
implements needed support for Sending messages from Browser to GPU process
over IPC. The implementation is a copy from OzoneDisplayChannelHost except
for the usage of IPC:: Channel. We rely on BrowserGpuChannelHostFactory to
establish the GPU Channel and BrowserChildProcessHostIterator to send
messages to correct process type. RemoteStateChangeHandler will be owned by
OzoneDisplayChannelHost. RemoteStateChangeHandler will be taken in to use
in follow up patch.