Add gypi file per folder.
This patch moves introduces a gypi file per folder. The gypi file
includes files in that folder, instead of having all of them in
ozone_impl.gyp. Now, Desktop Aura related files are compiled only
when toolkit_views = 1. This has resulted in a linking error when
Toolkit views is disabled but will be fixed in coming up patches.