WaylandDisplay: Make Destroy and Create functions private.
Any Widget state change needs to be handled via WindowStateChangeHandler
interface. We have CreateAcceleratedSurface and DestroyWindow which shouldn't
be used outside WaylandDisplay now. This CL makes these functions private.
Most of the changes are related to rearranging code in order to ensure the
order of declaration and implementation are same.