Wayland: StartProcessingEvents the first time a widget is realized.
Currently, OzoneDisplay informs WaylandDisplay to StartProcessingEvents when
the first AcceleratedWidget is Realized. WaylandDisplay however decides when
to StopProcessingEvents (i.e. When all surfaces are destroyed). Start/Stop
processing events can be dealt internally in WaylandDisplay as it owns the
poll thread and also it would simplify the logic in OzoneDisplay. This CL
changes so that WaylandDisplay starts Processing events the first time an
Accelerated surface is realized. Start/Stop processing functions will be
made private in a follow up patch.