IPC: Introduce OzoneChannelHostFactory.
Currently, OzoneDisplay is responsible for establishing the Channel
connection between ChannelHost and Channel. This should be transparent
in OzoneDisplay and should be handled in content. This patch introduces
OzoneChannelHostFactory which will be responsible for either establishing
IPC connection with DisplayChannel or just creating InProcessEventConverter,
depending on inprocess GPU usage. The idea here is that, we should be able
to establish the Browser process side connection in BrowserMainLoop as we
do in case of BrowserGpuChannelHostFactory.