Add support to provide external ozone files in events, gfx and views.
Ozone implementations for SFO, EventFactory and DesktopFactory should be
build as part of the respective target (i.e. gfx, events and views) and
Ozone should depend on them. Currently, this is not the case and we
build everything together into one wayland target. This patch adds
support for Ozone implementations to pass and build factory implementations
as part of the right target.