Stop requiring Weston as a build dependency.
This reverts the dependency on Weston 1.4.0 for building Ozone-Wayland
introduced in commit e57ab035f78d82117e1f1c7ec1459eb0c21deb22.

While it is OK to depend on Wayland and other libraries and even require
specific versions of them, it does not make much sense to do the same
for Weston, since it is not included or linked to by anything during the
build.