Bump Wayland version requirement to 1.5.0.
This is required at least for Tizen, which has recently moved from 1.4.0
to 1.5.0, and should also make the lives of most developers easier as
it's closer to the bleeding edge.

Since we now require a Wayland version newer than 1.4.x, a note has been
added to the README to mention XDG Shell support currently does not
work.