Relax Wayland version checks.
Instead of insisting Wayland to be at a specific version (1.4.0 at this
time of writing), make it a ">=" check.

This is necessary for Tizen, which has recently moved to Wayland 1.5.0,
and it also makes the lives of most developers easier as they can use a
version closer to the bleeding edge as well.

Since we now support Wayland versions newer than 1.4.x, a note has been
added to the README to mention XDG Shell support only works with the 1.4
series.