Get rid of ProxyDisplay.
We used proxydisplay for getting the right screen bounds on Browserprocess
side during the startup. ProxyDisplay inherited SFO and provided this
functionality, this resulted in additional complexity in case of
single process (Two SFO implementations). To avoid this, we initialized and
deleted the ProxyDisplay in single process mode. This CL gets rid of Proxy
Display completely and implements the same functionality in WaylandDisplay.