Check keyboard modifiers when returning cached symbol code.
Law keyboard code is converted to XKB code before sending
it to the browser process. At this time, the previous XKB code
is reused when the same law keyboard code is received.
However, "\" and "|" have the same keyboard code, so we distinguish
them with shift key(keyboard modifiers), but keyboard modifiers
have not been considered to reuse the previous XKB code.

This patch allows to check keyboard modifiers to distinguish
between "+" and "=", "\" and "|" when returning cached symbol code.

Bug: XWALK-1997