Define the window bounds as (0,0) when minimizing.
Propagate the window minimization information to the
content window, so the render side can update its
visibility properly. OnWMStateUpdated() is called by
PropertyNofify event from DispatchEvent() when the
browser is minimized or shown from minimized state.
On Windows, this is realized by calling OnHostResized()
with an empty size. Wayland on Linux should do it, too
(it just does not work on Linux with X11).

Bug: #295

Change-Id: If54d4f6468f97c12ce23c33c3d9934fdb5da815e
Author: Nicolas Guyomard <nicolas.guyomard@open.eurogiciel.org>
Signed-off-by: Manuel Bachmann <manuel.bachmann@open.eurogiciel.org>