DRWH: No need to pass size attributes as part of FullScreen, Restore.
In case of FullScreen and Restore, we handle resizing the surface in
DRWH. We also pass the size attributes as part of the IPC call to Display
and try to resize the surface again. The latter is not needed and we
can just let DRWH handle it. This patch makes changes in this front.