Don't set SwapInterval in case of OffScreen Surface.
We set EGLSwapInterval for both OnScreen and OffScreen surfaces.
In case of OffScreen surface, we use a Surfaceless Context and thus
we don't initialize any OffScreen EGL Surface. However, EGLSwapInterval
complains that EGL Surface is not present as it expects the draw buffers
to be present. This patch fixes the error by setting swap interval
only in case of OnScreen surface.