Initialize CursorData before Sandbox is enabled.
Initalizing Cursor Theme and loading images seem to result in system
calls which are not allowed under sandbox in GPU side. This Change
moves the initialation to happen during InitializeHardware call instead
of waiting till the Cursor is created. Eventually, we want to get rid of
all this and just use Cursor Images provided by Aura.