Wayland: Add experiment DRM support.
This patch adds experimental support to open a device, get the authentication
done through Wayland and use it to initialize GBM buffers. Idea is to use this
for Media later. Quick trial hangs my system, all the code is guarded with a compile
time flag and is disabled by default.