Do not generate and build content_common's libva stubs.
We are doing that ourselves in src/ozone/media/video.gypi, but we were
still generating and building the libva stubs in content_common.gypi as
well. This is fine on most systems since the code is not actually used,
but it depends on and includes X11 headers which might not exist on a
pure Wayland setup.