Resurrect CreateMediaOzonePlatformWayland(), sort of.
We still need the class as the build system in ui/ozone/ozone.gyp (the
generate_construct.py script) generates code that references this class,
and the build fails when linking in shared_library mode.

CreateMediaOzonePlatformWayland() now just returns a null pointer
though, as 0013-Media-Use-upstream-Video-Accelerator.patch changes the
code upstream not to need the function.

Related to #367.