Build the code in vaapi code in media/ as part of the content layer.
This fixes linking errors when building in shared_library mode, as the
code in vaapi_wrapper.cc needs access to Media::VideoFrame which is in
media/ and there is a circular dependency if we add the dependency in
the build system.

Instead, update
0005-Media-Build-VaapiPictureWayland-as-part-of-Media.patch and move the
code in media/video.gypi to content/content_common.gypi itself so we
build our code as part of the content layer.

Related to #367.