Always build the "wayland" target as a static library.
If we do not do that, we get several linking errors when building in
shared_library mode, as the code relies on several symbols from ui/ozone
itself that we cannot use if building as a separate, shared library. If
we always build as a static library, the symbols will be resolved when
libozone.so links agains libwayland.a and everything works.

Related to #367.