Add an error message when running with software rendering
If the GPU process is started in software rendering mode, it's almost
always because it has crashed too many times in hardware rendering mode.
Leave a helpful error message for new developers, since this commonly
occurs when building Ozone-Wayland for new devices or when hacking on
the GPU process, and it can take a while to figure out what's going on
if you haven't seen it before.