Wayland: Dont set the window as event grabber immediately after creation
We assume that newly created window has the focus. This is leading to
issues with popups, as we end up handling events which where meant for
the parent (i.e. When the popup is still not completely created). This
resulted in Menu getting closed as soon as it was created.

Fixes #390