Cursor: Add necessary support to use SKBitmap
This patch adds necessary support to pass PlatformCursor to GPU process
and use it as CursorBitmap. This patch also removes CursorFactory as
we no longer use it but rely on BitmapCursorFactoryOzone. There is
lot more cursor related code which can be removed and this will be
done in a follow up patch.

Fixes #343