Fix the crash problem happening during the input method initialization
When we set a instance of OzoneWebUI by calling LinuxUI::SetInstance in
ChromeBrowserMainExtraPartsAura::PreEarlyInitialization, the instance
should be initialized ahead by calling OzoneWebUI::Initialize(). However,
we didn't. As a result, it caused the crash in
InputMethodAuraLinux::InputMethodAuraLinux() because OzoneWebUI::Initialize()
was not called ChromeBrowserMainExtraPartsAura::ToolkitInitialized
due to the USE_X11 macro.

This CL allows OzoneWebUI::Initialize to be called in
ChromeBrowserMainExtraPartsAura::ToolkitInitialized.