Make DesktopFactoryOzone pick the proper DesktopFactoryOzone{Platform} on the fly
Patch does two main things:

1) Replaces 0002-Remove-usage-of-DesktopFactory.patch
by 0002-Adapt-DesktopFactoryOzone-to-respect-ozone-platform-.patch
in patches/. This way ozone/wayland continues to make use of
the DesktopFactoryOzone abstraction, and tweaks ::GetInstance to
create the proper DesktopFactoryOzone{Platform} instance on the fly.

2) Adds some GYP hooks that actually allows DesktopFactoryOzone::GetInstance
to instantiate the proper DesktopFactoryOzone{Platform} object
on demand, according to either --ozone-platform, or to a
default value.

TEST=regular GYP build
GYP_DEFINES="component=static_library
             clang=0
             use_sysroot=0
             linux_use_bundled_gold=0
             use_ozone=1
             ozone_auto_platforms=0
             ozone_platform_wayland=1
             use_xkbcommon=1"